requirements.txt
absl-py==2.1.0
aiohttp==3.10.10
attrs==23.2.0
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
grpcio==1.63.0
huggingface-hub==0.22.2
Pillow==10.0.0
python-telegram-bot==20.5
torch==2.0.1
transformers==4.34.0
groq==0.12.0
requests==2.32.3
setuptools>=44.0.0
uvicorn==0.29.0
fastapi==0.111.0
